split alternate operators combinators into separate module
