+++ examples/pyops.js
@@ -5,10 +5,10 @@ var fs = require('fs'),
     C = require('../lib/combinators'),
     O = require('../lib/operators');
 
-var pre = O.prefix3,
-    post = O.postfix3,
-    left = O.chainL3,
-    right = O.chainR3,
+var pre = O.prefix,
+    post = O.postfix,
+    left = O.chainL,
+    right = O.chainR,
     oneOf = C.position.oneOf,
     tok = C.position.literal,
     str = C.position.string,
