+++ test/operators.js
@@ -15,51 +15,11 @@ module('operators', function() {
         question = C.seq2R(C.basic.literal('?'), C.pure(function q(y) {return '(' + y + ')?';})),
         num = C.fmap(parseFloat, C.basic.oneOf('0123456789'));
     
-    test("chainL", function() {
-        var parser = O.chainL(plus, num),
-            a = parser.parse('8+4+2+1abc', 'state'),
-            v = a.value;
-        deepEqual(a.status, 'success');
-        deepEqual(v.rest, 'abc');
-        deepEqual(v.state, 'state');
-        deepEqual(v.result, [[[8,4],2],1]);
-    });
-    
-    test("chainL2", function() {
-        var parser = O.chainL2(plus, num),
-            a = parser.parse('8+4+2+1abc', 'state'),
-            v = a.value;
-        deepEqual(a.status, 'success');
-        deepEqual(v.rest, 'abc');
-        deepEqual(v.state, 'state');
-        deepEqual(v.result, [[[8,4],2],1]);
-    });
-    
     test("chainR", function() {
-        var parser = O.chainR(plus, num),
-            a = parser.parse('8+4+2+1abc', 'state'),
-            v = a.value;
-        deepEqual(a.status, 'success');
-        deepEqual(v.rest, 'abc');
-        deepEqual(v.state, 'state');
-        deepEqual(v.result, [8,[4,[2,1]]]);
-    });
-
-    test("chainR2", function() {
-        var parser = O.chainR2(plus, num),
-            a = parser.parse('8+4+2+1abc', 'state'),
-            v = a.value;
-        deepEqual(a.status, 'success');
-        deepEqual(v.rest, 'abc');
-        deepEqual(v.state, 'state');
-        deepEqual(v.result, [8,[4,[2,1]]]);
-    });
-    
-    test("chainR3", function() {
         var p = C.basic.literal('+'),
             n = C.basic.oneOf('0123456789'),
             util = require('util');
-        var parser = O.chainR3(p, n),
+        var parser = O.chainR(p, n),
             a = parser.parse('8+4+2+1abc', 'state'),
             v = a.value;
         deepEqual(a.status, 'success');
@@ -68,11 +28,11 @@ module('operators', function() {
         deepEqual(O.dump(v.result), '(+ 8 (+ 4 (+ 2 1)))');
     });
     
-    test("chainL3", function() {
+    test("chainL", function() {
         var p = C.basic.literal('+'),
             n = C.basic.oneOf('0123456789'),
             util = require('util');
-        var parser = O.chainL3(p, n),
+        var parser = O.chainL(p, n),
             a = parser.parse('8+4+2+1abc', 'state'),
             v = a.value;
         deepEqual(a.status, 'success');
@@ -82,47 +42,7 @@ module('operators', function() {
     });
     
     test("prefix", function() {
-        var parser = O.prefix(bang, num),
-            a = parser.parse('!!!8abc', 'state'),
-            v = a.value;
-        deepEqual(a.status, 'success');
-        deepEqual(v.rest, 'abc');
-        deepEqual(v.state, 'state');
-        deepEqual(v.result, '!(!(!(8)))');
-    });
-    
-    test("postfix", function() {
-        var parser = O.postfix(question, num),
-            a = parser.parse('8???abc', 'state'),
-            v = a.value;
-        deepEqual(a.status, 'success');
-        deepEqual(v.rest, 'abc');
-        deepEqual(v.state, 'state');
-        deepEqual(v.result, '(((8)?)?)?');        
-    });
-
-    test("prefix2", function() {
-        var parser = O.prefix2(bang, num),
-            a = parser.parse('!!!8abc', 'state'),
-            v = a.value;
-        deepEqual(a.status, 'success');
-        deepEqual(v.rest, 'abc');
-        deepEqual(v.state, 'state');
-        deepEqual(v.result, '!(!(!(8)))');
-    });
-    
-    test("postfix2", function() {
-        var parser = O.postfix2(question, num),
-            a = parser.parse('8???abc', 'state'),
-            v = a.value;
-        deepEqual(a.status, 'success');
-        deepEqual(v.rest, 'abc');
-        deepEqual(v.state, 'state');
-        deepEqual(v.result, '(((8)?)?)?');
-    });
-
-    test("prefix3", function() {
-        var parser = O.prefix3(C.basic.literal('!'), num),
+        var parser = O.prefix(C.basic.literal('!'), num),
             a = parser.parse('!!!8abc', 'state'),
             v = a.value;
         deepEqual(a.status, 'success');
@@ -131,8 +51,8 @@ module('operators', function() {
         deepEqual(O.dump(v.result), '(! (! (! 8)))');
     });
     
-    test("postfix3", function() {
-        var parser = O.postfix3(C.basic.literal('?'), num),
+    test("postfix", function() {
+        var parser = O.postfix(C.basic.literal('?'), num),
             a = parser.parse('8???abc', 'state'),
             v = a.value;
         deepEqual(a.status, 'success');
@@ -142,4 +62,3 @@ module('operators', function() {
     });
 
 });
-
